News-Crawler-staging/requirements.txt
brotlipy==0.7.0
colorama==0.4.6
datasets==2.14.4
google==3.0.0
google-cloud-pubsub==2.18.3
importlib-metadata==6.0.0
libcst==1.0.1
lxml==4.9.2
pymongo==4.5.0
pyopenssl==23.2.0
pysocks==1.7.1
python-dotenv==1.0.0
pyu2f==0.1.5
sacremoses==0.0.53
unidecode==1.3.6
newspaper3k==0.2.8
fastapi==0.101.1
gunicorn==21.2.0
python-multipart==0.0.6
uvicorn==0.23.2
dnspython==2.4.2
requests_html
